// src/app/common/enums/Alerts.ts
export enum Alerts {
  // Titulos de Alert
  T_SUCCESS = 'CORRECTO',
  T_INFO = 'INFORMACIÓN',
  T_WARNING = 'FALLO',
  T_ERROR = 'ERROR INESPERADO',

  // Mensage de Alert
  M_SUCCESS = 'Se ha ejecutado exitosamente',
  M_INFO = 'Los datos suministrados no son suficientes',
  M_WARNING = 'Algo salió mal!',
  M_ERROR = 'Ha ocurrido un error al comunicarse con el sistema',

  // Clases para visializar alerts
  SUCCESS = 'alert-success',
  INFO = 'alert-info',
  WARNING = 'alert-warning',
  ERROR = 'alert-danger',
}